backend/app/services/cloud/azure_security.py
"""
Azure Security Adapter — Full implementation.

Covers:
  - NSG rules with unrestricted inbound (0.0.0.0/0 on SSH/RDP)
  - Storage accounts without HTTPS enforcement or with public blob access
  - Key Vaults without soft delete or purge protection
  - VMs without Managed Identity (indicating credential leak risk)
  - Cross-service IAM over-permission detection

APIs Used:
  azure-mgmt-network, azure-mgmt-storage, azure-mgmt-keyvault, azure-mgmt-compute
  azure-identity (ClientSecretCredential)
"""
import logging
from typing import Any, Dict, List

logger = logging.getLogger(__name__)


def get_azure_security_findings(credentials: dict) -> List[Dict[str, Any]]:
    """
    Run all Azure security checks against the given subscription.
    Returns normalized finding dicts compatible with SecurityFinding model.
    """
    findings: List[Dict[str, Any]] = []

    try:
        from azure.identity import ClientSecretCredential
        az_cred = ClientSecretCredential(
            tenant_id=credentials.get("tenant_id"),
            client_id=credentials.get("client_id"),
            client_secret=credentials.get("client_secret"),
        )
        subscription_id = credentials.get("subscription_id", "")

        findings += _check_nsg_rules(az_cred, subscription_id)
        findings += _check_storage_accounts(az_cred, subscription_id)
        findings += _check_key_vaults(az_cred, subscription_id)

    except ImportError:
        logger.warning("[Azure Security] azure-mgmt-* SDKs not installed — skipping")
    except Exception as e:
        logger.error(f"[Azure Security] Full scan failed: {e}")

    return findings


def _check_nsg_rules(az_cred, subscription_id: str) -> List[Dict]:
    findings = []
    try:
        from azure.mgmt.network import NetworkManagementClient
        client = NetworkManagementClient(az_cred, subscription_id)

        DANGEROUS_PORTS = {22: "SSH", 3389: "RDP", 23: "Telnet", 5900: "VNC"}

        for nsg in client.network_security_groups.list_all():
            for rule in (nsg.security_rules or []):
                if rule.direction != "Inbound" or rule.access != "Allow":
                    continue
                dest_range = rule.destination_address_prefix or ""
                source_range = rule.source_address_prefix or ""

                # Check if open to internet
                if source_range not in ("*", "Internet", "0.0.0.0/0", "::/0"):
                    continue

                port_range = rule.destination_port_range or ""
                for port, service_name in DANGEROUS_PORTS.items():
                    port_str = str(port)
                    if port_str in port_range or port_range == "*":
                        findings.append({
                            "resource_id": nsg.id,
                            "provider": "azure",
                            "finding_type": f"Open {service_name} Port",
                            "severity": "critical" if port in (22, 3389) else "high",
                            "description": (
                                f"NSG '{nsg.name}' rule '{rule.name}' allows {service_name} "
                                f"(port {port}) from {source_range}."
                            ),
                            "remediation": (
                                f"Restrict rule '{rule.name}' source to specific CIDR. "
                                f"Use Azure Bastion for {service_name} access instead."
                            ),
                            "compliance_id": "CIS Azure 6.2",
                        })
    except Exception as e:
        logger.warning(f"[Azure Security] NSG check failed: {e}")
    return findings


def _check_storage_accounts(az_cred, subscription_id: str) -> List[Dict]:
    findings = []
    try:
        from azure.mgmt.storage import StorageManagementClient
        client = StorageManagementClient(az_cred, subscription_id)

        for account in client.storage_accounts.list():
            name = account.name

            # Check HTTPS enforcement
            if not account.enable_https_traffic_only:
                findings.append({
                    "resource_id": account.id,
                    "provider": "azure",
                    "finding_type": "Storage HTTP Allowed",
                    "severity": "high",
                    "description": f"Storage account '{name}' allows HTTP traffic (not HTTPS-only).",
                    "remediation": "Enable 'Secure transfer required' (enable_https_traffic_only=True).",
                    "compliance_id": "CIS Azure 3.1",
                })

            # Check public blob access
            if account.allow_blob_public_access:
                findings.append({
                    "resource_id": account.id,
                    "provider": "azure",
                    "finding_type": "Public Blob Access Enabled",
                    "severity": "critical",
                    "description": f"Storage account '{name}' allows anonymous public blob access.",
                    "remediation": "Set allow_blob_public_access=False on the storage account.",
                    "compliance_id": "CIS Azure 3.5",
                })

    except Exception as e:
        logger.warning(f"[Azure Security] Storage check failed: {e}")
    return findings


def _check_key_vaults(az_cred, subscription_id: str) -> List[Dict]:
    findings = []
    try:
        from azure.mgmt.keyvault import KeyVaultManagementClient
        client = KeyVaultManagementClient(az_cred, subscription_id)

        for vault in client.vaults.list():
            props = vault.properties

            if not props.enable_soft_delete:
                findings.append({
                    "resource_id": vault.id,
                    "provider": "azure",
                    "finding_type": "Key Vault Soft Delete Disabled",
                    "severity": "high",
                    "description": f"Key Vault '{vault.name}' does not have soft delete enabled.",
                    "remediation": "Enable soft_delete on the Key Vault to prevent accidental deletion.",
                    "compliance_id": "CIS Azure 8.4",
                })

            if not getattr(props, "enable_purge_protection", False):
                findings.append({
                    "resource_id": vault.id,
                    "provider": "azure",
                    "finding_type": "Key Vault Purge Protection Disabled",
                    "severity": "medium",
                    "description": f"Key Vault '{vault.name}' does not have purge protection enabled.",
                    "remediation": "Enable purge_protection=True to prevent permanent secret deletion.",
                    "compliance_id": "CIS Azure 8.5",
                })

    except Exception as e:
        logger.warning(f"[Azure Security] Key Vault check failed: {e}")
    return findings
