requirements.txt
# ── Core API Framework ────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
uvloop==0.19.0
httptools==0.6.1
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==23.2.1
slowapi==0.1.9
httpx==0.27.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.31
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.2

# ── Cache + PubSub ────────────────────────────────────────────────────────────
redis==5.0.7

# ── Authentication ────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
cryptography==42.0.8

# ── Data Validation ───────────────────────────────────────────────────────────
pydantic==2.8.2
pydantic-settings==2.3.4
email-validator==2.2.0

# ── Cloud SDKs ────────────────────────────────────────────────────────────────
boto3==1.34.144
botocore==1.34.144
google-cloud-compute>=1.20.0
google-cloud-monitoring>=2.22.0
google-auth==2.31.0
google-api-python-client==2.137.0
google-cloud-billing>=1.18.0
google-cloud-securitycenter>=1.30.0
azure-mgmt-compute==31.0.0
azure-mgmt-monitor==6.0.2
azure-mgmt-resource==23.1.1
azure-identity==1.17.1
azure-mgmt-consumption==9.0.0
azure-mgmt-costmanagement==4.0.1
azure-mgmt-security==6.0.0

# ── Event Streaming ───────────────────────────────────────────────────────────
confluent-kafka==2.4.0
aiokafka==0.11.0
faststream[kafka]==0.5.24

# ── ML ────────────────────────────────────────────────────────────────────────
scikit-learn==1.5.1
numpy==1.26.4
scipy==1.14.0
joblib==1.4.2

# ── Graph ─────────────────────────────────────────────────────────────────────
networkx==3.3

# ── Kubernetes ────────────────────────────────────────────────────────────────
kubernetes==30.1.0

# ── Observability ─────────────────────────────────────────────────────────────
setuptools==70.3.0
# Let google-cloud-sdk determine protobuf version
protobuf>=4.21.0,<6.0.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-sqlalchemy==0.46b0
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==7.0.0

# ── System Monitoring ─────────────────────────────────────────────────────────
psutil==6.0.0

# ── AI & Vault ────────────────────────────────────────────────────────────────
hvac==2.3.0
google-generativeai==0.7.2

tenacity==8.2.3
